+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 0.2.0 (Unreleased)
+
+### Features Added
+
+- Scheduling messages to be delivered at a later date, via the `Sender.ScheduleMessage(s)` function or 
+  setting `Message.ScheduledEnqueueTime`.
+- Added in the `Sender.SendMessages([slice of sendable messages])` function, which batches messages 
+  automatically. Useful when you're sending multiple messages that you are already sure will be small
+  enough to fit into a single batch.
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 0.1.0 (2021-10-05)
 
 - Initial preview for the new version of the Azure Service Bus Go SDK.